requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
jinja2